[BE] feat: 추천 API 개발 (#458)
* [BE] fix: 채택률 응답 형식 정수형으로 수정 #454

* [BE] fix: 채택률 스케줄러 cron 표현식 적용 #454

* [BE] feat: 페르소나별 추천 응답에 프리뷰 이미지 및 추천 옵션 추가 #454

* [BE] feat: 연령대에 따른 채택률 상위 1위 외장 및 내장 색상 응답 추가 #454

* [BE] feat: 라이프스타일별 추천 응답에 추천 색상 및 옵션 추가 #454

+++ backend/src/main/java/com/softeer/caart/domain/color/dto/ColorDto.java
@@ -32,8 +32,8 @@ public class ColorDto {
 		private final Long colorId;
 		private final String colorName;
 		private final String colorImage;
-		private final Integer colorPrice;
-		private final double adoptionRate;
+		private final int colorPrice;
+		private final int adoptionRate;
 		private final List<String> previews;
 
 		public ExteriorColorDto(Color color, double adoptionRate) {
@@ -41,7 +41,7 @@ public class ColorDto {
 			this.colorName = color.getName();
 			this.colorImage = color.getImage().getUrl();
 			this.colorPrice = color.getPrice();
-			this.adoptionRate = adoptionRate;
+			this.adoptionRate = (int)adoptionRate;
 			this.previews = color.getImageUrlListOfColorPreview();
 		}
 	}
@@ -51,8 +51,8 @@ public class ColorDto {
 		private final Long colorId;
 		private final String colorName;
 		private final String colorImage;
-		private final Integer colorPrice;
-		private final double adoptionRate;
+		private final int colorPrice;
+		private final int adoptionRate;
 		private final String preview;
 
 		public InteriorColorDto(Color color, double adoptionRate) {
@@ -60,7 +60,7 @@ public class ColorDto {
 			this.colorName = color.getName();
 			this.colorImage = color.getImage().getUrl();
 			this.colorPrice = color.getPrice();
-			this.adoptionRate = adoptionRate;
+			this.adoptionRate = (int)adoptionRate;
 			this.preview = color.getImageUrlOfMainColorPreview();
 		}
 	}
